+++ service/internal/transport/rest/dto/case.go
@@ -8,7 +8,7 @@ type Case struct {
 	Camera         Camera    `json:"camera,omitempty"`
 	Violation      Violation `json:"violation,omitempty"`
 	ViolationValue string    `json:"violation_value,omitempty"`
-	RequiredSkill  int       `json:"required_skill,omitempty"`
+	RequiredSkill  int64     `json:"required_skill,omitempty"`
 	Date           time.Time `json:"date,omitempty"`
 	IsSolved       bool      `json:"is_solved,omitempty"`
 	FineDecision   bool      `json:"fine_decision,omitempty"`
